--- resource-manager/managementgroups/2023-04-01/managementgroups/method_ubscriptionsgetsubscriptionsundermanagementgroup.go
@@ -1,106 +1,0 @@
-package managementgroups
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-
-	"github.com/hashicorp/go-azure-helpers/resourcemanager/commonids"
-	"github.com/hashicorp/go-azure-sdk/sdk/client"
-	"github.com/hashicorp/go-azure-sdk/sdk/odata"
-)
-
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See NOTICE.txt in the project root for license information.
-
-type UbscriptionsGetSubscriptionsUnderManagementGroupOperationResponse struct {
-	HttpResponse *http.Response
-	OData        *odata.OData
-	Model        *[]SubscriptionUnderManagementGroup
-}
-
-type UbscriptionsGetSubscriptionsUnderManagementGroupCompleteResult struct {
-	LatestHttpResponse *http.Response
-	Items              []SubscriptionUnderManagementGroup
-}
-
-type UbscriptionsGetSubscriptionsUnderManagementGroupCustomPager struct {
-	NextLink *odata.Link `json:"nextLink"`
-}
-
-func (p *UbscriptionsGetSubscriptionsUnderManagementGroupCustomPager) NextPageLink() *odata.Link {
-	defer func() {
-		p.NextLink = nil
-	}()
-
-	return p.NextLink
-}
-
-// UbscriptionsGetSubscriptionsUnderManagementGroup ...
-func (c ManagementGroupsClient) UbscriptionsGetSubscriptionsUnderManagementGroup(ctx context.Context, id commonids.ManagementGroupId) (result UbscriptionsGetSubscriptionsUnderManagementGroupOperationResponse, err error) {
-	opts := client.RequestOptions{
-		ContentType: "application/json; charset=utf-8",
-		ExpectedStatusCodes: []int{
-			http.StatusOK,
-		},
-		HttpMethod: http.MethodGet,
-		Pager:      &UbscriptionsGetSubscriptionsUnderManagementGroupCustomPager{},
-		Path:       fmt.Sprintf("%s/subscriptions", id.ID()),
-	}
-
-	req, err := c.Client.NewRequest(ctx, opts)
-	if err != nil {
-		return
-	}
-
-	var resp *client.Response
-	resp, err = req.ExecutePaged(ctx)
-	if resp != nil {
-		result.OData = resp.OData
-		result.HttpResponse = resp.Response
-	}
-	if err != nil {
-		return
-	}
-
-	var values struct {
-		Values *[]SubscriptionUnderManagementGroup `json:"value"`
-	}
-	if err = resp.Unmarshal(&values); err != nil {
-		return
-	}
-
-	result.Model = values.Values
-
-	return
-}
-
-// UbscriptionsGetSubscriptionsUnderManagementGroupComplete retrieves all the results into a single object
-func (c ManagementGroupsClient) UbscriptionsGetSubscriptionsUnderManagementGroupComplete(ctx context.Context, id commonids.ManagementGroupId) (UbscriptionsGetSubscriptionsUnderManagementGroupCompleteResult, error) {
-	return c.UbscriptionsGetSubscriptionsUnderManagementGroupCompleteMatchingPredicate(ctx, id, SubscriptionUnderManagementGroupOperationPredicate{})
-}
-
-// UbscriptionsGetSubscriptionsUnderManagementGroupCompleteMatchingPredicate retrieves all the results and then applies the predicate
-func (c ManagementGroupsClient) UbscriptionsGetSubscriptionsUnderManagementGroupCompleteMatchingPredicate(ctx context.Context, id commonids.ManagementGroupId, predicate SubscriptionUnderManagementGroupOperationPredicate) (result UbscriptionsGetSubscriptionsUnderManagementGroupCompleteResult, err error) {
-	items := make([]SubscriptionUnderManagementGroup, 0)
-
-	resp, err := c.UbscriptionsGetSubscriptionsUnderManagementGroup(ctx, id)
-	if err != nil {
-		result.LatestHttpResponse = resp.HttpResponse
-		err = fmt.Errorf("loading results: %+v", err)
-		return
-	}
-	if resp.Model != nil {
-		for _, v := range *resp.Model {
-			if predicate.Matches(v) {
-				items = append(items, v)
-			}
-		}
-	}
-
-	result = UbscriptionsGetSubscriptionsUnderManagementGroupCompleteResult{
-		LatestHttpResponse: resp.HttpResponse,
-		Items:              items,
-	}
-	return
-}
